Feat: 예약 관련 서비스 코드 작성

+++ src/main/java/com/mycom/myapp/domain/reservation/dto/ReservationDto.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ReservationDto {
 	
-	private int roomId;
+	private long roomId;
 	private LocalDateTime reservationDate;
 	private LocalDateTime startDate;
 	private LocalDateTime endDate;
